Create CODE_OF_CONDUCT.md
Hi guys, this is the code of conduct directly from Contributor Covenant.   Please give feedback as to if you would like be to add the team lead and mentor to be points of call prior to calling buildforsdg@andela.com. 

We could also adopt the code and refer to the Contributor Covenant -  I did do an adjusted version, but if it is good enough for 200 000 open source projects I think it should suffice for us. :)